+++ ui/packages/app/grafana-panel-plugin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.8](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.7...@parca/grafana-panel@0.0.8) (2022-11-02)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
 ## [0.0.7](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.6...@parca/grafana-panel@0.0.7) (2022-11-02)
 
 **Note:** Version bump only for package @parca/grafana-panel